+++ SPAJAM2025/PublisherFiles/PublisherGameView.swift
@@ -35,25 +35,4 @@ enum PublisherViewIdentifier: String {
     PublisherGameView()
 }
 
-struct StarHidingView: View {
-    @Binding var currentView: PublisherViewIdentifier
-    var body: some View {
-        VStack{
-            Spacer()
-            Text("あなたの星を")
-                .font(.title)
-            Text("夜空にまぎれさせましょう")
-                .font(.title)
-            
-            Spacer()
-            
-            Button{
-                currentView = .waitingForQuestion
-            }label: {
-                Text("これで決定！")
-                    .padding()
-            }
-            .buttonStyle(.borderedProminent)
-        }
-    }
-}
+
