+++ SPAJAM2025/ReceiverFiles/ReceiverFindingView.swift
@@ -1,0 +1,16 @@
+//
+//  ReceiverFindingView.swift
+//  SPAJAM2025
+//
+//  Created by Kota Yamaguchi on 2025/09/27.
+//
+
+import SwiftUI
+
+struct ReceiverFindingView: View {
+    var body: some View {
+        VStack{
+            
+        }
+    }
+}
