+++ SPAJAM2025/ReceiverFiles/ReceiverQuestionView.swift
@@ -1,0 +1,16 @@
+//
+//  ReceiverQuestionView.swift
+//  SPAJAM2025
+//
+//  Created by Kota Yamaguchi on 2025/09/27.
+//
+
+import SwiftUI
+
+struct ReceiverQuestionView: View {
+    var body: some View {
+        VStack{
+            
+        }
+    }
+}
