+++ SPAJAM2025/ReceiverFiles/ReceiverResultView.swift
@@ -1,0 +1,16 @@
+//
+//  ReceiverResultView.swift
+//  SPAJAM2025
+//
+//  Created by Kota Yamaguchi on 2025/09/27.
+//
+
+import SwiftUI
+
+struct ReceiverResultView: View {
+    var body: some View {
+        VStack{
+            
+        }
+    }
+}
